% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{alph}
\alias{alph}
\title{Extract subject level summary stats from a sampler}
\usage{
alph(sampler, func = median)
}
\arguments{
\item{sampler}{A pmwgs sampler containing mcce samples}

\item{func}{A summary stat function}
}
\value{
A tibble
}
\description{
This function extracts subject level summary statistics from a pmwgs object
}
