% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{arch_medians}
\alias{arch_medians}
\title{Reshape median architecture estimates for plotting}
\usage{
arch_medians(medians, grouping_vars = c("subjectid"))
}
\arguments{
\item{medians}{}
}
\value{
A tibble augmented ready for plotting
}
\description{
Reshape median architecture estimates for plotting
}
