% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{arch_plot}
\alias{arch_plot}
\title{Plot an architecture comparison barplot}
\usage{
arch_plot(medians)
}
\arguments{
\item{medians}{The medians for each architecture/person}
}
\value{
A ggplot object
}
\description{
...
}
