% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{check_env}
\alias{check_env}
\title{Check the system environment variables and extract the relevant pieces for
modelling}
\usage{
check_env()
}
\value{
A list with the elements extracted from the sys env
}
\description{
Check the system environment variables and extract the relevant pieces for
modelling
}
