% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dirichlet_reduced_mix}
\alias{dirichlet_reduced_mix}
\title{Top level log-likelihood function that implements reduced dirichlet sampling}
\usage{
dirichlet_reduced_mix(
  x,
  data,
  contaminant_prob = 0.02,
  alpha_indices = c(1, 2),
  min_rt = 0,
  max_rt = 1,
  shared = FALSE
)
}
\arguments{
\item{x}{A named vector containing parameter values to test}

\item{data}{The data for a single subject for which the likelihood should be
calculated}

\item{contaminant_prob}{The probability used for contaminant process in the
modelling. A contaminant process is just a uniform random response in the
allowable time window.}

\item{alpha_indices}{A vector containing the indicies of the alpha parameters
\itemize{
\item that is the parameters that correspond to the dirichlet shape parameters.
}}

\item{min_rt}{The smallest possible response time in the data}

\item{max_rt}{The largest possible response time in the data}
}
\value{
The log of the likelihood for the data under parameter values x
}
\description{
This function selects one of the possible architectures using a random
dirichlet sample weighted by the α values from the parameter vector. It
then delegates the likelihood for the data to an architecture specific
log-likelihood function.
}
\section{The parameter vector}{


The vector x should contain the following elements:
A number of \eqn{\alpha}
values

\itemize{
\item A number of \strong{\eqn{\alpha}} parameter values matching the
number and order of the ll_funcs vector defined in this same file.
\item \strong{A} - the start point variability
\item \strong{b^a} and \strong{b^r}, the thresholds to either accept or
reject the item.
\item \strong{t0} - the residual time, bounded above by the minimum
response time for the participant
\item 12 drift rates. For each attribute there are three stimulus levels.
For each of these 6 attribute levels there are two drift rates, one drift
rate to accept (\strong{v^a}) and one to reject (\strong{v^r})
}
}

