% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{extract_alphas}
\alias{extract_alphas}
\title{Extract alpha parameters (architectures) from the samples}
\usage{
extract_alphas(sampler)
}
\arguments{
\item{sampler}{The pmwgs sampler object}
}
\value{
A tibble with the alpha parameter samples and a subjectid column
}
\description{
This function taks a pmwgs sampler object and extracts both the group level
and the individual subject level samples for the alpha parameter estimates.
}
