% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{extract_data}
\alias{extract_data}
\title{Use read_csv and bind_rows to rearrange data from JATOS file}
\usage{
extract_data(raw_data)
}
\arguments{
\item{raw_data}{The data in raw format to extract from}
}
\value{
All experimental data as a tibble
}
\description{
Use read_csv and bind_rows to rearrange data from JATOS file
}
