% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{extract_tmu}
\alias{extract_tmu}
\title{Extract group level parameters from the samples}
\usage{
extract_tmu(
  sampler,
  par_names = sampler$par_names,
  filter = unique(sampler$samples$stage)
)
}
\arguments{
\item{sampler}{The pmwgs sampler object}

\item{par_names}{The names of the parameters to extract - defaults to all
parameters estimated.}

\item{filter}{The sampling stages to extract - defaults to all stages run.
Can also be a vector of sample indices.}
}
\value{
A tibble with the parameter samples and a sampleid column
}
\description{
This function taks a pmwgs sampler object and extracts the group level
samples for the specified parameter estimates.
}
