% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for one architecture for all subjects}
\usage{
generate_data(model_data, vars)
}
\arguments{
\item{model_data}{The original data for the task being simulated.}

\item{vars}{The list of job variables.}

\item{medians}{The median estimates of parameter values for each participant.}
}
\value{
A tibble with the simulated data values matching the original data.
}
\description{
This function takes some data, median parameter estimates and job variables
and outputs a tibble with simulated data for all subjects
}
