% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_base_filename}
\alias{get_base_filename}
\title{Create an output basename from envars list}
\usage{
get_base_filename(vars)
}
\arguments{
\item{vars}{The list of object extracted from the environment}
}
\value{
a character vector with 1 element, the basename
}
\description{
From a list extracted from environment variable for a modelling job
extract the necessary element for a basename for output files.
}
