% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{get_drifts}
\alias{get_drifts}
\title{Takes a sample tibble and rearranges it to a long df with just drift rates.}
\usage{
get_drifts(sample_df)
}
\arguments{
\item{sample_df}{A dataframe with columns for each parameter and sampleid}
}
\value{
A long tibble containing the drift rate type and values
}
\description{
Takes a sample tibble and rearranges it to a long df with just drift rates.
}
