% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_estimation_data}
\alias{get_estimation_data}
\title{Extract the data from a sampler object stored in an RData file}
\usage{
get_estimation_data(filename)
}
\arguments{
\item{filename}{The RData file containing the sampler object}
}
\value{
The data tibble that was used as input to the modelling
}
\description{
The sampler object is assumed to be called "sampler", no checks are done
}
