% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_JATOS_data}
\alias{get_JATOS_data}
\title{Get data from a raw JSON file}
\usage{
get_JATOS_data(datafile)
}
\arguments{
\item{datafile}{A string containing the local of a JATOS data export}
}
\value{
A list containing each participants data converted from JSON
}
\description{
Reads data from a JATOS results file.
Each line from the file should be a JSON object from the PrefSFT2020 code
}
