% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{get_medians}
\alias{get_medians}
\title{Return median values for  parameters (from extract_parameters)}
\usage{
get_medians(pars, alpha = TRUE)
}
\arguments{
\item{pars}{The result of extract_parameters, tibble containing samples for
the selected parameter estimates.}
}
\value{
A tibble containing the medians of the samples for each subject
}
\description{
Return median values for  parameters (from extract_parameters)
}
