% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_samples}
\alias{get_samples}
\title{Load an RData file with pmwg samples for further analysis}
\usage{
get_samples(pmwg_file, final_obj = "sampled")
}
\arguments{
\item{pmwg_file}{The environment from a run of pmwg as RData}

\item{final_obj}{The object (generally a pmwgs object) to extract}
}
\value{
The pmwgs object
}
\description{
Load an RData file with pmwg samples for further analysis
}
