% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{get_summary}
\alias{get_summary}
\title{Return summary values for parameters (from extract_parameters)}
\usage{
get_summary(pars, tform = base::identity, sfunc = stats::median)
}
\arguments{
\item{pars}{The result of extract_parameters, tibble containing samples for
the selected parameter estimates.}

\item{tform}{Any transformation necessary to the data prior to summarising}

\item{sfunc}{A summary function (one value), default is median}
}
\value{
A tibble containing the medians of the samples for each subject
}
\description{
Return summary values for parameters (from extract_parameters)
}
