% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{known_vars}
\alias{known_vars}
\title{Known environment variables}
\format{
An object of class \code{character} of length 18.
}
\usage{
known_vars
}
\description{
Known environment variables
}
\keyword{datasets}
