% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{ll_CNST}
\alias{ll_CNST}
\title{Coactive No/Self terminating Yes}
\usage{
ll_CNST(rt, A, b_pos, b_neg, t0, v_pos, v_neg, positive)
}
\arguments{
\item{rt}{A vector of response times}

\item{A}{Start point variability}

\item{b_pos}{positive evidence threshold}

\item{b_neg}{negative evidence threshold}

\item{t0}{non decision time parameter}

\item{v_pos}{positive evidence drift rate (vector of)}

\item{v_neg}{vector of negative evidence drift rates}

\item{positive}{Whether we are looking at accept or reject trials}
}
\value{
The log likelihood of the rts for the accept or reject trials given
the provided parameter values
}
\description{
Coactive No/Self terminating Yes
}
