% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{ll_CYST}
\alias{ll_CYST}
\title{Coactive Yes Self terminating No}
\usage{
ll_CYST(rt, A, b_acc, b_rej, t0, v_pos, v_neg, accept)
}
\arguments{
\item{rt}{A vector of response times}

\item{A}{Start point variability}

\item{b_acc}{positive evidence threshold}

\item{b_rej}{negative evidence threshold}

\item{t0}{non decision time parameter}

\item{v_pos}{positive evidence drift rate (vector of)}

\item{v_neg}{vector of negative evidence drift rates}

\item{accept}{Whether we are looking at accept or reject trials}
}
\value{
The log likelihood of the rts for the accept or reject trials given
the provided parameter values
}
\description{
Coactive Yes Self terminating No
}
