% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architectures.R, R/old_loglike.R
\name{ll_FPP}
\alias{ll_FPP}
\title{First Past the Post}
\usage{
ll_FPP(rt, A, b_acc, b_rej, t0, drifts, accept)

ll_FPP(rt, A, b_acc, b_rej, t0, drifts, accept)
}
\arguments{
\item{rt}{A vector of response times}

\item{A}{Start point variability}

\item{b_acc}{positive evidence threshold}

\item{b_rej}{negative evidence threshold}

\item{t0}{non decision time parameter}

\item{drifts}{An array of drift rates, 4 x length(rt) with the columns being
vectors of drift rates for accept price, reject price, accept rating,
reject rating respectively}

\item{accept}{Whether we are looking at accept or reject trials}
}
\value{
The log likelihood of the rts for the accept or reject trials given
the provided parameter values

The log likelihood of the rts for the accept or reject trials given
the provided parameter values
}
\description{
First Past the Post

First Past the Post
}
