% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{matdf_plot}
\alias{matdf_plot}
\title{Plot matrix_to_df object using ggplot}
\usage{
matdf_plot(x)
}
\arguments{
\item{x}{A data.frame containing an X and Y column for parameter labels
and a value column containing the data to plot.}
}
\value{
A ggplot object
}
\description{
This function takes a matrix in data.frame form, usually output from
matrix_to_df and displays it using ggplot.
The resulting figure can be adjusted further.
}
