% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Reshape a correlation matrix into a data.frame}
\usage{
matrix_to_df(x)
}
\arguments{
\item{x}{A matrix with named x and y dimensions.}
}
\value{
A df with X, Y and value columns
}
\description{
This function takes a square matrix, like a correlation matrix derived from a
mean covariance matrix, and reshapes it into a data.frame for plotting.
}
