% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcce-package.R
\docType{package}
\name{mcce-package}
\alias{mcce}
\alias{mcce-package}
\title{mcce: Models of Consumer Choice Experiments}
\description{
This package contains the likelihood functions and scripts necessary to recreate the modelling work for my PhD.
}
\author{
\strong{Maintainer}: Gavin Cooper \email{gavin@gavincooper.net} (\href{https://orcid.org/0000-0002-7186-5293}{ORCID})

}
\keyword{internal}
