% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_loglike.R, R/wrappers.R
\name{model_wrapper}
\alias{model_wrapper}
\title{Wrapper for individual model log likelihood function}
\usage{
model_wrapper(x, data, model)

model_wrapper(x, data, model)
}
\arguments{
\item{x}{A named vector containing parameter values to test}

\item{data}{The data for a single subject for which the likelihood should be
calculated}

\item{model}{The model to be wrapped and returned}
}
\value{
The log of the likelihood for the data under parameter values x

The log of the likelihood for the data under parameter values x
}
\description{
This function performs some common steps such as rearrangeing the data
pulling out parameter items into variable and combining the results of
applying to model to different subsets of the data (accept and reject
responses)

This function performs some common steps such as rearrangeing the data
pulling out parameter items into variable and combining the results of
applying to model to different subsets of the data (accept and reject
responses)
}
