% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{order_arch_medians}
\alias{order_arch_medians}
\title{Reorder architectures and subjects of medians for plotting}
\usage{
order_arch_medians(model_medians, arch_order = NULL)
}
\arguments{
\item{model_medians}{}
}
\value{
ordered model medians
}
\description{
From a arch_medians tibble, reorder the subjectID's and architectures
in order for them to be presented in a plot nicely.
}
