% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{parse_JATOS_line}
\alias{parse_JATOS_line}
\title{Parses a JATOS datafile line in 2020 Pref SFT format}
\usage{
parse_JATOS_line(data_line, line_index)
}
\arguments{
\item{data_line}{The string containing the line in question}

\item{line_index}{The line number of the line being processed}
}
\value{
The result of parsing the JSON from the line, or an error
}
\description{
For a particular line from a datafile, parse the JSON and return the result.
}
