% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot subject and group level summaries of estimates}
\usage{
plot_summary(alpha_summ, tmu_summ, transform = identity)
}
\arguments{
\item{alpha_summ}{A data.frame containing subject level summary stats of
each parameter}

\item{tmu_summ}{A data.frame containing group level summary stats}

\item{transform}{An optional function to transform the summary stats.}
}
\value{
A ggplot object
}
\description{
This function takes subject level and group level summary statistics
from an mcce sampling run and creates a plot showing the results.
}
