% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{print_JATOS_line}
\alias{print_JATOS_line}
\title{Print a problem JATOS datafile line if JSON parsing failed}
\usage{
print_JATOS_line(data_line, line_index)
}
\arguments{
\item{data_line}{The string containing the line in question}

\item{line_index}{The line number of the line being processed}
}
\description{
Print a problem JATOS datafile line if JSON parsing failed
}
