% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_expyriment_data}
\alias{read_expyriment_data}
\title{Read data generated by the Expyriment python library}
\usage{
read_expyriment_data(folder, filename_pattern)
}
\arguments{
\item{folder}{The data folder (string)}

\item{filename_pattern}{The pattern with which the names of each data file
start (string)}
}
\description{
Import Exypriment data into R. The function concatinates all data and returns
a R data frame with all subjects. Between subject factors will be added as
variables to the data matrix.
}
\details{
Copyright: 2012-2015 Florian Krause \href{mailto:siebenhundertzehn@googlemail.com}{siebenhundertzehn@googlemail.com}
2012-2015 Oliver Lindemann \href{mailto:lindemann09@googlemail.com}{lindemann09@googlemail.com}
License: GPL-3.0+
Adapted for this package by Gavin Cooper \href{mailto:gavin@gavincooper.net}{gavin@gavincooper.net} 2020
}
