% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{reformat}
\alias{reformat}
\title{Reformat data from the Expyriment file format to a useful tibble}
\usage{
reformat(x)
}
\arguments{
\item{folder}{The data folder (string)}

\item{filename_pattern}{The pattern with which the names of each data file
start (string)}
}
\value{
A list containing each participants data converted from JSON
}
\description{
Import Exypriment data into R. The function concatinates all data and returns
a R data frame with all subjects. Between subject factors will be added as
variables to the data matrix.
}
\details{
Copyright: 2012-2015 Florian Krause \href{mailto:siebenhundertzehn@googlemail.com}{siebenhundertzehn@googlemail.com}
2012-2015 Oliver Lindemann \href{mailto:lindemann09@googlemail.com}{lindemann09@googlemail.com}
License: GPL-3.0+
Adapted for this package by Gavin Cooper \href{mailto:gavin@gavincooper.net}{gavin@gavincooper.net} 2020
}
