% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{rll_MW}
\alias{rll_MW}
\title{Max Winner random samples}
\usage{
rll_MW(data, A, b_acc, b_rej, t0, drifts)
}
\arguments{
\item{data}{the data for one subject}

\item{A}{Start point variability}

\item{b_acc}{positive evidence threshold}

\item{b_rej}{negative evidence threshold}

\item{t0}{non decision time parameter}

\item{drifts}{An array of drift rates, 4 x length(rt) with the columns being
vectors of drift rates for accept price, reject price, accept rating,
reject rating respectively}
}
\value{
A new data object with the same shape and new randomly drawn
responses and RT's.
}
\description{
Max Winner random samples
}
