% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architectures.R
\name{select_ll}
\alias{select_ll}
\title{Select log likelihood function}
\usage{
select_ll(name, sample = FALSE)
}
\arguments{
\item{name}{The name of the log likelihpood function to select}

\item{sample}{If TRUE then return the sample function instead of the density
function}
}
\value{
the requested function
}
\description{
Select log likelihood function
}
