% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{simple_model_wrapper}
\alias{simple_model_wrapper}
\title{Simple wrapper for integrate testing}
\usage{
simple_model_wrapper(rt, A, b_acc, b_rej, t0, drifts, accept, model)
}
\arguments{
\item{rt}{A vector of response times}

\item{A}{Start point variability}

\item{b_acc}{positive evidence threshold}

\item{b_rej}{negative evidence threshold}

\item{t0}{non decision time parameter}

\item{drifts}{An array of drift rates, 4 x length(rt) with the columns being
vectors of drift rates for accept price, reject price, accept rating,
reject rating respectively}

\item{accept}{Whether we are looking at accept or reject trials}

\item{model}{The loglikelihood function currently being tested}
}
\value{
The log of the likelihood for the rt's for the parameters from the
model
}
\description{
This function does the parameter transformations necessary for the model to
run.
}
\section{The parameter vector}{


The vector x should contain the following elements:
A number of \eqn{\alpha}
values

\itemize{
\item \strong{A} - the start point variability
\item \strong{b^a} and \strong{b^r}, the thresholds to either accept or
reject the item.
\item \strong{t0} - the residual time, bounded above by the minimum
response time for the participant
\item 12 drift rates. For each attribute there are three stimulus levels.
For each of these 6 attribute levels there are two drift rates, one drift
rate to accept (\strong{v^a}) and one to reject (\strong{v^r})
}
}

