% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{tmu}
\alias{tmu}
\title{Extract group level summary stats from a sampler}
\usage{
tmu(sampler, func = median)
}
\arguments{
\item{sampler}{A pmwgs sampler containing mcce samples}

\item{func}{A summary stat function}
}
\value{
A tibble
}
\description{
This function extracts group level summary statistics from a pmwgs object
}
