% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot a trace for a subject/parameter combination}
\usage{
trace_plot(sample_df, par)
}
\arguments{
\item{sample_df}{The tibble with samples.}

\item{par}{The parameter to plot a trace of}
}
\value{
A ggplot object
}
\description{
For a particular sample tibble (from extract_parameters) take one
parameter and plot the trace of the parameter value. The tibble should
be filtered to reduce the subjects, or faceted to separate the subjects.
}
