% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{transform_pars}
\alias{transform_pars}
\title{Transforms parameters for specific models/}
\usage{
transform_pars(pars, data, tforms = c("std"))
}
\arguments{
\item{pars}{The named list of parameters to transform}

\item{tforms}{A character vector selecting the transformation scheme
scheme to be used.}
}
\value{
A named list of transformed parameter values
}
\description{
This function performs common transforms for the LBA based architecture of
choice modelling
}
\section{Transformation schemes}{


Transformations scheme currently allowed are provided below.
\itemize{
\item std - This is the default, and takes the exponent of all
parameter values, enforces \strong{b_acc} and \strong{b_rej} being
greater than \strong{A} by adding \strong{A}
}

