% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{get_medians}
\alias{get_medians}
\title{Return median values for alpha parameters (from extract_alphas)}
\usage{
get_medians(alphas)
}
\arguments{
\item{alphas}{The result of extract_alphas, tibble containing samples for
the architecture parameter estimates.}
}
\value{
A tibble containing the medians of the samples for each subject as
well as the relative weight for each architecture.
}
\description{
Return median values for alpha parameters (from extract_alphas)
}
