% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architectures.R, R/old_loglike.R
\name{names_ll}
\alias{names_ll}
\title{Return names of all implemented models (log likelihood functions)}
\usage{
names_ll()

names_ll()
}
\value{
names of the ll_funcs package object

names of the ll_funcs package object
}
\description{
Return names of all implemented models (log likelihood functions)

Return names of all implemented models (log likelihood functions)
}
