% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot subject and group level summaries of estimates}
\usage{
plot_summary(summ, transform = identity)
}
\arguments{
\item{summ}{A data.frame containing subject and group level summary stats
for each parameter. The data.frame should have a subjectid column with
theta_mu representing the group level estimates.}

\item{transform}{An optional function to transform the summary stats.}
}
\value{
A ggplot object
}
\description{
This function takes subject level and group level summary statistics
from an mcce sampling run and creates a plot showing the results.
}
