% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{rll_IST}
\alias{rll_IST}
\title{Independant Self Terminating random samples}
\usage{
rll_IST(x, data)
}
\arguments{
\item{x}{The parameter vector}

\item{data}{the data for one subject}
}
\value{
A new data object with the same shape and new randomly drawn
responses and RT's.
}
\description{
Independant Self Terminating random samples
}
