% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{transform_pars}
\alias{transform_pars}
\title{Transforms parameters for specific models/}
\usage{
transform_pars(pars, tforms = "std")
}
\arguments{
\item{pars}{The named list of parameters to transform}

\item{tforms}{A character vector selecting the transformation scheme
scheme to be used.}
}
\value{
A named list of transformed parameter values
}
\description{
This function performs common transforms for the LBA based architecture of
choice modelling
}
\section{Transformation schemes}{


Transformations scheme currently allowed are provided below.
\itemize{
\item std - This is the default, and takes the exponent of all
parameter values, enforces \strong{b_acc} and \strong{b_rej} being
greater than \strong{A} by adding \strong{A}
\item reduced - This works similarly to std, except it does not take the
exponent of drift rates (parameters starting with v). It also calculates
drift rate parameters for reject accumulators from the accept drift rates
combined with a beta_0 and beta_1 parameters for each attribute.
\item reduced_more - This works similarly to reduced, however the beta_0
parameters is shared between attributes.
}
}

